Fixing github actions
The fact that the package needs SpinFRGLattices,
that is at the moment not registered,
and also some sub packages,
makes setting up the testing environment
a little tricky.

For example, when testing PMFRGSolve,
we need to have SpinFRGLattices available
(and we can Pkg.add it),
but when we try to add it
it will complain that PMFRGCore is not registered,
Adding PMFRGCore as a dev depenency
trigger the opposite problem:
an error message saying that SpinFRGLattices is not registered.

The solution is to add both packages as dev dependencies
at the same time,
using Pkg.develop
with a vector of PackageSpec objects.