ExObsUpdate: proj/deps load conditional on version
The project/deps objects
that are created at the beginning
of the automatically generated ExampleObservable.jl file
belong to the Pkg API
which is not very stable,
so if they are read by different julia version
they might let to a crash of the test suite.

The representation of the proj/deps objects
has been sandwiched in an if statement
that allows the objects to be created
only if the julia VERSION
is the one used when generating the values.

Also, the code has been refactored in smaller functions.