Fix uuid, version, authorship in Project.toml
When doing the split of the package
I assumed that PMFRG would become PMFRGCore
So PMFRGCore "stole" version and authorship from PMFRG,
and PMFRG got new uuid, version and authorship.
and PMFRGCore got a new uuid too.

uuid of PMFRGCore has also been fixed
in Project.toml of dependants (PMFRGSolve and PMFRG).

This has been reverted now.
(This is done so that it is possible to try and register the packages
in a local registry.)